+++ test/unittests/state_encoding.c
@@ -210,7 +210,7 @@ UNIT(EncodeInvertsDecode_Base2)
 {
     uint64_t const base = 2;
     uint64_t const M = 10;
-    for (uint64_t x = 0; x < (uint64_t)pow(base,M); ++x)
+    for (uint64_t x = 0; x < (uint64_t) pow((double) base, (double) M); ++x)
     {
        uint64_t *state = inform_decode(x,M,base);
        ASSERT_EQUAL_U(x, inform_encode(state,M,base));
@@ -222,7 +222,7 @@ UNIT(EncodeInvertsDecode_Base3)
 {
     uint64_t const base = 3;
     uint64_t const M = 6;
-    for (uint64_t x = 0; x < (uint64_t)pow(base,M); ++x)
+    for (uint64_t x = 0; x < (uint64_t) pow((double) base, (double) M); ++x)
     {
        uint64_t *state = inform_decode(x,M,base);
        ASSERT_EQUAL_U(x, inform_encode(state,M,base));
@@ -234,7 +234,7 @@ UNIT(EncodeInvertsDecode_Base4)
 {
     uint64_t const base = 4;
     uint64_t const M = 5;
-    for (uint64_t x = 0; x < (uint64_t)pow(base,M); ++x)
+    for (uint64_t x = 0; x < (uint64_t) pow((double) base, (double) M); ++x)
     {
        uint64_t *state = inform_decode(x,M,base);
        ASSERT_EQUAL_U(x, inform_encode(state,M,base));
@@ -246,7 +246,7 @@ UNIT(EncodeInvertsDecode_Base5)
 {
     uint64_t const base = 5;
     uint64_t const M = 4;
-    for (uint64_t x = 0; x < (uint64_t)pow(base,M); ++x)
+    for (uint64_t x = 0; x < (uint64_t) pow((double) base, (double) M); ++x)
     {
        uint64_t *state = inform_decode(x,M,base);
        ASSERT_EQUAL_U(x, inform_encode(state,M,base));
